% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal_calendar_convention_india_public.R
\name{calendar_quarter}
\alias{calendar_quarter}
\title{Calendar quarter extractor}
\usage{
calendar_quarter(x, with_year = TRUE, auto_convert_fiscal_quarter = TRUE)
}
\arguments{
\item{x}{A character vector or Date vector}

\item{with_year}{Logical; whether to include year (Qx:YYYY) or only Qx}
}
\value{
A character vector of calendar quarters
}
\description{
Extract calendar quarter labels from text or Date inputs.
}
\examples{
calendar_quarter("GDP grew in Q2:2025")
calendar_quarter(as.Date("2025-08-15"))
calendar_quarter(as.Date("2025-08-15"), with_year = FALSE)

}
