% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_readers.R
\name{es_convert}
\alias{es_convert}
\title{Convert e-Sankhyiki GDP/GVA Data to TDF Long Format}
\usage{
es_convert(dat)
}
\arguments{
\item{dat}{A data frame containing the raw e-Sankhyiki data.}
}
\value{
An object of class \code{tdf_long}.
}
\description{
Transforms raw annual or quarterly e-Sankhyiki data (such as Gross Value Added
or Gross Domestic Product) into a standardized \code{tdf_long} object.
}
