% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_with_attrs.R
\name{tibble_with_attrs}
\alias{tibble_with_attrs}
\alias{as_tibble_with_attrs}
\title{Create a tibble that preserves attributes}
\usage{
tibble_with_attrs(x, ...)

as_tibble_with_attrs(x)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{...}{Name-value pairs of attributes to attach.}
}
\value{
A \code{tibble_with_attrs}.
}
\description{
Create a tibble that preserves attributes
}
\examples{
d <- tibble_with_attrs(iris, source = "datasets", built = Sys.time())
attr(dplyr::filter(d, Sepal.Length > 5), "source")
}
