% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal_calendar_convention_india_public.R
\name{as.Date.fiscal_period}
\alias{as.Date.fiscal_period}
\title{Coerce fiscal_period to Date}
\usage{
\method{as.Date}{fiscal_period}(x, anchor = c("last", "first", "mid"), ...)
}
\arguments{
\item{x}{A character vector of class \code{fiscal_period}.}

\item{anchor}{One of "first", "mid", or "last".}

\item{...}{Unused.}
}
\value{
A Date vector of the same length as \code{x}.
}
\description{
Converts a \code{fiscal_period} object to a Date using
the specified anchor.
}
